Baseball/requirements.txt
appdirs==1.4.3
Brotli==1.0.9
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
dash==1.15.0
dash-core-components==1.11.0
dash-daq==0.5.0
dash-html-components==1.1.0
dash-renderer==1.7.0
dash-table==4.10.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.1
packaging==20.3
pandas==1.1.1
pep517==0.8.2
Pillow==7.2.0
plotly==4.9.0
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
